requirements.txt
langchain-core
google-generativeai
langchain_google_genai